refactor a little bit
